# Speed up BoundaryStats by preparing boundary polygon